+++ R/helpers.R
@@ -258,10 +258,14 @@ setMethod(
 #' @return
 #' See \code{\link[base]{all.equal}}
 all.equal.simList <- function(target, current, ...) {
-  attr(target, "newCache") <- NULL
-  attr(current, "newCache") <- NULL
+  attr(target, ".Cache")$newCache <- NULL
+  attr(current, ".Cache")$newCache <- NULL
   attr(target, "removedObjs") <- NULL
   attr(current, "removedObjs") <- NULL
+  suppressWarnings(rm("._startClockTime", envir = envir(target)))
+  suppressWarnings(rm("._startClockTime", envir = envir(current)))
+  suppressWarnings(rm(".timestamp", envir = envir(target)))
+  suppressWarnings(rm(".timestamp", envir = envir(current)))
 
   # make a dummy class so it uses the default all.equal
   all.equal.default(target, current)