+++ R/simulation-simInit.R
@@ -342,6 +342,12 @@ setMethod(
 
     names(modules) <- unlist(modules)
 
+    # Check that modules exist in paths$modulePath
+    moduleDirsExist <- dir.exists(file.path(paths$modulePath, unlist(modules)))
+    if (!isTRUE(all(moduleDirsExist))) {
+      stop("These modules: ", unlist(modules)[!moduleDirsExist], " , don't exist in ", paths$modulePath)
+    }
+
     # identify childModules, recursively
     childModules <- .identifyChildModules(sim = sim, modules = modules)
     modules <- as.list(unique(unlist(childModules))) # flat list of all modules
@@ -367,7 +373,12 @@ setMethod(
     reqdPkgs <- packages(modules=unlist(modules), paths = paths(sim)$modulePath,
                          envir = sim@.envir[[".parsedFiles"]])
     if (length(unlist(reqdPkgs))) {
-      Require(c(unique(unlist(reqdPkgs), "SpaDES.core")))
+      allPkgs <- c(unique(unlist(reqdPkgs), "SpaDES.core"))
+      if (getOption("spades.useRequire")) {
+        Require(allPkgs)
+      } else {
+        loadedPkgs <- lapply(allPkgs, require, character.only = TRUE)
+      }
     }
 
     ## timeunit is needed before all parsing of modules.