+++ inst/examples/example_moduleMetadata.R
@@ -1,5 +1,9 @@
 path <- system.file("sampleModules", package = "SpaDES.core")
 sampleModules <- dir(path)
+# turn off code checking -- don't need it here
+opts <- options("spades.moduleCodeChecks" = FALSE,
+                "spades.useRequire" = FALSE)
+
 x <- moduleMetadata(sampleModules[3], path = path)
 
 # using simList
@@ -12,3 +16,6 @@ mySim <- simInit(
    paths = list(modulePath = system.file("sampleModules", package = "SpaDES.core"))
 )
 moduleMetadata(sim = mySim)
+
+# turn code checking back on -- don't need it here
+options(opts)