+++ tests/testthat/test-load.R
@@ -1,10 +1,14 @@
 test_that("test-load.R: loading inputs does not work correctly", {
-  library(igraph)
-  tmpdir <- file.path(tempdir(), "test_load", rndstr()) %>% checkPath(create = TRUE)
+  testInitOut <- testInit()
   on.exit({
-    detach("package:igraph")
-    unlink(tmpdir, recursive = TRUE)
+    testOnExit(testInitOut)
   }, add = TRUE)
+  # library(igraph)
+  # tmpdir <- file.path(tempdir(), "test_load", rndstr()) %>% checkPath(create = TRUE)
+  # on.exit({
+  #   detach("package:igraph")
+  #   unlink(tmpdir, recursive = TRUE)
+  # }, add = TRUE)
 
   mapPath <- system.file("maps", package = "quickPlot")
 
@@ -113,9 +117,9 @@ test_that("test-load.R: loading inputs does not work correctly", {
 })
 
 test_that("test-load.R: passing arguments to filelist in simInit does not work correctly", {
-  tmpdir <- file.path(tempdir(), "test_load", rndstr()) %>% checkPath(create = TRUE)
+  testInitOut <- testInit()
   on.exit({
-    unlink(tmpdir, recursive = TRUE)
+    testOnExit(testInitOut)
   }, add = TRUE)
 
   # Second, more sophisticated. All maps loaded at time = 0, and the last one is reloaded
@@ -204,11 +208,9 @@ test_that("test-load.R: passing arguments to filelist in simInit does not work c
 })
 
 test_that("test-load.R: passing objects to simInit does not work correctly", {
-  library(igraph)
-  tmpdir <- file.path(tempdir(), "test_load", rndstr()) %>% checkPath(create = TRUE)
+  testInitOut <- testInit()
   on.exit({
-    detach("package:igraph")
-    unlink(tmpdir, recursive = TRUE)
+    testOnExit(testInitOut)
   }, add = TRUE)
 
   mapPath <- mapPath <- system.file("maps", package = "quickPlot")
@@ -281,6 +283,11 @@ test_that("test-load.R: passing objects to simInit does not work correctly", {
 })
 
 test_that("test-load.R: passing nearly empty file to simInit does not work correctly", {
+  testInitOut <- testInit()
+  on.exit({
+    testOnExit(testInitOut)
+  }, add = TRUE)
+
   mapPath <- system.file("maps", package = "quickPlot")
 
   # test object passing directly
@@ -306,10 +313,9 @@ test_that("test-load.R: passing nearly empty file to simInit does not work corre
 })
 
 test_that("test-load.R: more tests", {
-  tmpdir <- file.path(tempdir(), "test_load", rndstr()) %>% checkPath(create = TRUE)
-
+  testInitOut <- testInit()
   on.exit({
-    unlink(tmpdir, recursive = TRUE)
+    testOnExit(testInitOut)
   }, add = TRUE)
 
   sim <- simInit()
@@ -350,13 +356,9 @@ test_that("test-load.R: more tests", {
 })
 
 test_that("test-load.R: interval loading of objects from .GlobalEnv", {
-  library(igraph)
-  tmpdir <- file.path(tempdir(), "test_load", rndstr()) %>% checkPath(create = TRUE)
-  cwd <- getwd()
-  setwd(tmpdir)
+  testInitOut <- testInit()
   on.exit({
-    unlink(tmpdir, recursive = TRUE)
-    setwd(cwd)
+    testOnExit(testInitOut)
   }, add = TRUE)
 
   times <- 0:10